[#52] Extract cypher property structure
