+++ src/translator/cypher/graph/CMakeLists.txt
@@ -2,6 +2,7 @@ add_library(scc_translator_cypher_graph STATIC)
 target_sources(scc_translator_cypher_graph PRIVATE
         constraint_types.cpp
         node.cpp
+        property.cpp
         property_types.cpp
         relationship.cpp
 )
