+++ src/translator/cypher/graph/node.cpp
@@ -17,24 +17,6 @@ bool operator<(const Label& lhs, const Label& rhs) {
   return lhs.value < rhs.value;
 }
 
-Property::Property(std::string  name, std::string value, PropertyType type)
-    : name(std::move(name)), value(std::move(value)), type(type) {}
-
-std::string Property::ToString() const {
-  return name + ": " + value;
-}
-
-std::ostream& operator<<(std::ostream& os, const Property &property) {
-  os << property.ToString();
-  return os;
-}
-bool operator<(const Property& lhs, const Property& rhs) {
-  if (lhs.type != rhs.type) {
-    return lhs.type < rhs.type;
-  }
-  return lhs.name < rhs.name;
-}
-
 Node::Node(std::string label_name)
     : labels({Label(std::move(label_name))}) {}
 Node::Node(std::string label_name, std::string variable)
